refactor name of subpackage from tools to utils
