+++ deepsearcher/agent/collection_router.py
@@ -2,7 +2,7 @@ from typing import List, Tuple
 
 from deepsearcher.agent.base import BaseAgent
 from deepsearcher.llm.base import BaseLLM
-from deepsearcher.tools import log
+from deepsearcher.utils import log
 from deepsearcher.vector_db.base import BaseVectorDB
 
 COLLECTION_ROUTE_PROMPT = """
